requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
Django==4.0.4
django-phonenumber-field==6.1.0
iniconfig==1.1.1
packaging==21.3
phonenumbers==8.12.49
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
